EL-3430 - Timeline Chart Build Issue
