+++ docs/app/pages/charts/charts-sections/timeline-chart/timeline-chart.module.ts
@@ -1,6 +1,6 @@
 import { ComponentFactoryResolver, NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { ColorServiceModule, HybridModule, TabsetModule } from '@ux-aspects/ux-aspects';
+import { ColorServiceModule, HybridModule, TabsetModule, TimelineChartModule } from '@ux-aspects/ux-aspects';
 import { ChartsModule } from 'ng2-charts';
 import { DocumentationComponentsModule } from '../../../../components/components.module';
 import { DocumentationCategoryComponent } from '../../../../components/documentation-category/documentation-category.component';
@@ -32,6 +32,7 @@ const ROUTES = [
         HybridModule,
         RouterModule.forChild(ROUTES),
         TabsetModule,
+        TimelineChartModule,
         WrappersModule,
     ],
     exports: SECTIONS,
