+++ docs/app/pages/charts/charts-sections/timeline-chart/timeline-chart/timeline-chart.component.html
@@ -27,10 +27,7 @@
     needs to be added to the appropriate <code>NgModule</code>.
 </p>
 <p>
-    To add Timeline functionality the <code>TimelineChartPlugin</code> should be imported from
-    the <code>@ux-aspects/ux-aspects</code> library and the <code>TimelineChartPlugin.register()</code> function should be called.
-    This function registers the plugin globally and must be called before the chart initializes. It can be called in the consuming
-    component constructor, or if used by multiple components it may also be called within a module.
+    To add Timeline functionality the <code>TimelineChartModule</code> should be imported from the <code>@ux-aspects/ux-aspects</code> library.
 </p>
 
 <p>The a <code>timeline</code> object should be added to the chart options to configure the chart:</p>
